auditoria/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dotenv==0.9.9
et_xmlfile==2.0.0
fastapi==0.116.1
google-auth==2.40.3
google-genai==1.27.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.3
openpyxl==3.1.5
pandas
pdfkit==1.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic
pydantic_core
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
tenacity==8.5.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
